+++ packages/utils/src/node.ts
@@ -1,13 +1,4 @@
 /**
- * Checks whether we're in the Node.js or Browser environment
- *
- * @returns Answer to given question
- */
-export function isNodeEnv(): boolean {
-  return Object.prototype.toString.call(typeof process !== 'undefined' ? process : 0) === '[object process]';
-}
-
-/**
  * Requires a module which is protected against bundler minification.
  *
  * @param request The module path to resolve
